Update datatable.css
Added a style rule that applies to the debug status located in the navbar at the top right of the page. Currently this status box is not aligned properly vertically. I've added a line height property to extend the box size to match the size of the search box that is located to its right.

I also added margins to properly align it with its neighboring containers.