Refactored Versal Visualization Using Closure
First step to fixing multi-graph bug was to add in closure to separate
scope on multiple graphs. Also I wrapped each circle in a graphics tag
and this seemed to fix the force drag rendering with multiple graphs.
Still there is an issue with the second graph not being rendered.